Initial attempt at linux XPCOM component

+++ components/media/MediaRecorder.h
@@ -81,7 +81,7 @@
 #define SAMPLE          PRInt16
 #define SAMPLE_RATE     22050
 #define SAMPLE_FORMAT   paInt16
-#define SAMPLE_QUALITY  0.4
+#define SAMPLE_QUALITY  (float)(0.4)
 
 typedef struct {
     ogg_page og;
